fix(mojo-0263): resolve Mojo 0.26.3 syntax deprecations and add gradient checker traits
Fix all CI failures caused by Mojo 0.26.3 language changes:

Syntax deprecations (PR1 equivalent):
- Replace fn with def in function type parameter positions (4 files)
- Remove unified {read} from struct instance methods — only valid on
  nested functions (tests/conftest.mojo, test_lenet5_e2e.mojo,
  test_base_loader.mojo)
- Fix unused loop variables: for i/step/batch -> for _ where unused
  (test_mxfp4_block.mojo 18 loops, test_nvfp4_block.mojo 16 loops,
  test_list_append_stress.mojo, test_sgd.mojo, test_confusion_matrix_bugs.mojo,
  test_mixed_precision.mojo)
- Qualify implicit stdlib imports with std. prefix (test_model_utils.mojo,
  test_serialization.mojo, test_io_helpers.mojo, test_all_architectures.mojo,
  test_lenet5_e2e.mojo)
- Hoist from python import Python above try: blocks to function scope
  (test_model_utils.mojo, test_serialization.mojo, test_serialization utils)
- Add explicit Float64() casts for deprecated implicit Int->Float64 conversions
  (test_mixed_precision_simd.mojo)
- Add mypy type annotations: total_changes: dict[str, int] = {} (2 scripts)
- Rename std alias to stdev to avoid collision with std stdlib package
  (test_reduction.mojo)
- Replace String.substr() with slice syntax (test_serialization utils)

Gradient checker trait infrastructure (PR2 equivalent):
- Add NumericalBackward trait to gradient_checker.mojo (Copyable + Movable,
  __call__(grad_out, x) -> AnyTensor)
- Add check_gradients[F: NumericalForward, B: NumericalBackward] overload
- Add check_gradients_verbose[F: NumericalForward, B: NumericalBackward] overload
- Add check_gradient[F: NumericalForward, B: NumericalBackward] overload
- Export NumericalForward and NumericalBackward from shared/testing/__init__.mojo

All existing bare-function-type overloads remain unchanged (backward compatible).
Test file closure -> struct migration follows in subsequent PRs.

+++ tests/shared/training/test_mixed_precision_simd.mojo
@@ -92,7 +92,7 @@ def test_convert_fp16_to_fp32_large() raises:
 
     # Fill with incrementing values
     for i in range(1024 * 1024):
-        fp16_ptr[i] = Float16(Float32((i % 100) / 100.0))
+        fp16_ptr[i] = Float16(Float32(Float64(i % 100) / 100.0))
 
     # Convert to FP32
     var fp32_result = convert_to_fp32_master(fp16_params)
@@ -100,7 +100,7 @@ def test_convert_fp16_to_fp32_large() raises:
     # Spot check values
     var fp32_ptr = fp32_result._data.bitcast[Float32]()
     for i in [0, 100, 1000, 10000, 100000, 1000000 - 1]:
-        var expected = Float32((i % 100) / 100.0)
+        var expected = Float32(Float64(i % 100) / 100.0)
         var actual = fp32_ptr[i]
         assert_true(
             abs(actual - expected) < 0.001,
@@ -192,7 +192,7 @@ def test_convert_fp32_to_fp16_large() raises:
 
     # Fill with incrementing values
     for i in range(1024 * 1024):
-        fp32_ptr[i] = Float32((i % 100) / 100.0)
+        fp32_ptr[i] = Float32(Float64(i % 100) / 100.0)
 
     # Create FP16 target
     var fp16_model = AnyTensor([1024, 1024], DType.float16)
@@ -203,7 +203,7 @@ def test_convert_fp32_to_fp16_large() raises:
     # Spot check values
     var fp16_ptr = fp16_model._data.bitcast[Float16]()
     for i in [0, 100, 1000, 10000, 100000, 1000000 - 1]:
-        var expected = Float16(Float32((i % 100) / 100.0))
+        var expected = Float16(Float32(Float64(i % 100) / 100.0))
         var actual = fp16_ptr[i]
         assert_true(
             abs(Float32(actual) - Float32(expected)) < 0.001,
